pyproject.toml
[project]
name = "refined-claude"
version = "0.1"
description = "Accessibility refinements to Claude Desktop"
readme = "README.md"
authors = [{name = "Edward Z. Yang"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "click",
    "pyobjc-core==10.3.2",
    "pyobjc-framework-applicationservices==10.3.2",
    "pyobjc-framework-cocoa==10.3.2",
    "pyobjc-framework-quartz==10.3.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ruff>=0.11.0",
]

[project.urls]
Homepage = "https://github.com/ezyang/refined-claude"
Changelog = "https://github.com/ezyang/refined-claude/releases"
Issues = "https://github.com/ezyang/refined-claude/issues"
CI = "https://github.com/ezyang/refined-claude/actions"

[project.scripts]
refined-claude = "refined_claude.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
